[analyzer] Fix missing analyzer binary
If one of the registered analyzer binary (e.g: clang-tidy) was missing
from user host machine, in CodeChecker we tried to get the full binary
path from the user's `PATH` environment variable for this binary by using
the `find_executable` function but it returned with `None` and we tried
to get the `realpath` of it but the parameter of this function cannot be
`None` so we got an exception.

This commit will solve this issue by checking the return value of the
`find_executable` function and if it returned with `None` we will
print a warning message to the user.