[analyzer] Fix failure zip
In the analysis manager we used `tu_collector` to create failed zip
but after it added source files to the zip we opened this zip file
again in write mode to add some files but it cleared the zip so
the collected source files were removed from the zip.

To solve this problem we need to open the zip file in append mode.