[pylint] Fix undefined variable to thrift types
- We run pylint on analyzer/web directory separately and we give the
same `--rcfile` option to both command. This configuration file contains an
`init-hook` setting where we add multiple directories to the `sys.path`.
Unfortunatelly these paths will be added relative to the directory where
the pylint command will be executed. To solve this problem we will set the
`PYLINTRC` environment variable instead of using the `--rcfile` option. In
the init-hook we can get this variable, get the directory where this file
can be found and add absolute paths instead of relatives to the sys.path.
- Add `thrift` target as a dependency of pylint command in web module.
- Fix undefined variable errors reported by pylint.