less argument processing if skip file used without ctu and stats
if a skip file is used but no pre analysis step is executed
the pre analysis skip handler will not be used.

In the analyze.py the statistics collector flag
status was only checked if the ctu_phases was in the args.
The two of them should be checked independently because
stats and ctu can be executed independently.
stats can be enabled even if ctu is not enabled.

Counting of the skipped compilation needs to be done
at two places at the compile command parsing phase and during the
analysis phase too. Not everyting is skipped and the compile
command parsing phase because it might be needed at a pre analysis
step.