[analyzer] Only disable explicitly disabled checkers.
The handling of checkers and checks are modified. The analyzer decides which
modules are ran, and this implicitly handled state has to be modelled inside
CodeChecker in order to maintain the ability to list all existing
checks in the web module as well. This modification aims not to change
the detection state behaviour, explicitly enabled checks/checkers are
emitted as enabled, default and explicitly disabled ones end up being
emitted as disabled.