store cppcheck plist reports (generate hash)
cppcheck can generate plist reports but no bug identifier
hash is generated. A 0 value is generated by cppcheck as a bug hash.
This change allows to generate the report hash if the plist report
contains onl results where all the hashes are 0.

The generated report hash will be written to the plist report files.

Test if the hash is generated for the cppcheck plist reports.