[analyzer] Fix plist file name collision
Source files with equal filename generates and overwrites the
filename.plist result file during analysis step. The main problem
is that the filename hash generation contains only the filename and
not the file path.
This commit will solve this problem and include the buildaction
directory (removes the users home directory from the path) to the
filename hash.