[gui] Fix delete run
The server side run sort feature changed the Thrift API and introduced a
new paramter for `getRunData` function. In the `ListOfRuns.js` file
we forgot to change this function call in one place so the browser
is thrown the following exception:
codeCheckerDBAccess.js:102 Uncaught TypeError: this.sortMode.write is not a function

This commit fix the `getRunData` function call in the GUI's code.