[fix] Fix hiding tooltip of the source component filter item
If we set a new source component filter and we move the mouse over
this item a tooltip will be shown. If I remove this filter the tooltip
will not be removed and it will be visible constantly.

This commit solves this problem and hides the tooltip when the source
component filter item is being removed.