[analyzer][feature] Detect OSX clang version
Clang compiler shipped by Apple uses a different name in the version
string. This is now detected, and version info is extracted on OSX
platform as well.

Some gcc/g++ compiler flag filtering tests needed to be turned off
if gcc/g++ is just a symlink to clang/clang++, because different
filtering needs to be done for the compilers.

Co-authored-by: Endre Fulop <gamesh411@gmail.com>