[plist2html] Print total number of reports in plist-to-html
Running `CodeChecker parse -e html reports -o html_reports` will convert
the reports to html but the `Total number of reports: 0` even if there
were reports. This commit will fix this and will print the number of reports
and summary header in the source code of the `plist-to-html` tool and not in
the CodeChecker parse command.