create pypi and npm packages from the generated thrift api files
The new directory structure and build targets enable
to create and publish pypi and npm (jquery, node) packages from the
generated codechecker api thrift IDL's.

- new documentation was added to describe the API change workflow
- the new published packages were added ad external py and js dependecies
- cleanup travis CI thrift deps, docs, make targets
- bump up api package version to v6.25.0, version increment was needed
because publishing the pypi packages with wrong dependencies, with the
newer version it is fixed now. The thrift dependency needs to be a fixed
version (==0.9.2).