[fix][report-converter] Fix finding module in report-converter
If we run report-converter (cli.py) in an environment where
'codechecker_report_converter' module is not available we should add
the grandparent directory of the cli.py file to the system path. So
python will find the module correctly.

This problem is possible when we call report-converter from the build
directory of CodeChecker: `build/CodeChecker/bin/report-converter`.