[fix] Fix changing detection status to OFF
There was a problem with CodeChecker when .clang-tidy configration
file is used. If we enabled a checker in this config file and we
do not enabled it explicitly in CodeChecker if we stored the results
to the server these reports are marked as OFF.

With this patch if we store a report directory to the server then a
bug which can be found in this report directory can not be marked as
OFF so the detection status will be calculated properly.