[cmd] Improve readability of argparse help messages
We would like to improve the readability of the arg parser help messages
by allowing to write new line characters to the text but
`RawDescriptionDefaultHelpFormatter` will not keep white spaces.

For this reason we override the default behaviour of the `_split_lines`
method. If the text parameter starts with 'R|' we will keep whitespaces and
we will wrapp the content. Otherwise it will work the same as
RawDescriptionHelpFormatter.