[gui] Set limit and query when getting filter items
There was a bug in the UI when we set a file path filter for example
to `*` and we reloaded the page. Unfortunately we forgot to set the limit
and query parameters so it tried to get all the items from the server
which can be really large and a browser can be freezed.
This commit will solve this problem.