[server] Fix printing product statuses
There is a problem with the server when we do the following:
- Start a web server by using postgresql database.
- Add a new product by using sqlite database.
- Remove the sqlite database.
- Restart the server.
- The following exception will be thrown:
```
output_formatters.py", line 114, in twodim_to_table
    str_parts.append(print_string.format(*line))
TypeError: unsupported format string passed to NoneType.__format__
```

The problem at here that the schema version will be None and we try to
print it in a format string: "{0:0}.format(schema_ver)". To solve this
problem we will convert the schema version to string before print.