[server] Convert ldap search values to text
I got an error when I tried to run a CodeChecker server with a
postgres database, using psycopg2 driver and ldap configration.

```
  File "/codechecker/lib/python3/codechecker_server/permissions.py", line 217, in has_permission
    groups = self._has_perm_impl(auth_session.groups, True)
  File "/codechecker/lib/python3/codechecker_server/permissions.py", line 434, in _has_perm_impl
    exists = self.__session.query(query.exists()).scalar()
  ...
  sqlalchemy.exc.ProgrammingError: (psycopg2.ProgrammingError) operator does not exist: text = bytea
  LINE 3: ...roduct_id = 5 AND lower(permissions_product.name) IN ('\x6b6...
                                                             ^
  HINT:  No operator matches the given name and argument type(s). You might need to add explicit type casts.
```
The problem is that when we moved to python3 we forgot to converter
bytes to string when we getting group names from the return value of
`connection.search_s` function.

The LDAP protocol states that some fields be encoded in UTF-8. Attribute
values, on the other hand, MAY contain any type of data, including text.
Encoding/decoding to other formats – text, images, etc. – is left to the
caller. See:
https://www.python-ldap.org/en/python-ldap-3.2.0/bytes_mode.html

This commit will solve this problem.