[server] Handle database deadlocks when storing reports
When we use multiple server instances and we try to run multiple storage
to each server which contain at least two same reports which have source
code comments it is possible that the following exception will be thrown:
  (psycopg2.extensions.TransactionRollbackError) deadlock detected

The problem is that the report hash is the key for the review data table
and both of the store actions try to update the same review data row.

Neither of the two processes can continue, and they will wait for each other
indefinitely. PostgreSQL in this case will terminate one transaction with
the above exception. For this reason in case of failure we will wait some
seconds and try to run the storage again.

For more information see #2655 and #2653 issues on github.