[analyzer] Compiler errors as reports can be disabled.
In an earlier commit (#2652) a new feature has been introduced: compiler
errors can be captured as CodeChecker reports. The corresponding checker
name is clang-diagnostic-error. This type of report couldn't have been
disabled. In this commit these can be disabled too.

Warnings can be enabled and disabled as CodeChecker reports as well. For
example the warning about unused variables can be disabled by
-d Wunused-variable. This warning is captured as
clang-diagnostic-unused-variable and now this can be also disabled as
-d clang-diagnostic-unused-variable.