[tu-collector] Improve logging
The logger in the tu_collector was not initialized properly. This way
a lot of message was printed twice to the output. For example:
```
[INFO 2020-04-15 11:56] - [4/4] clangsa analyzed lib.cpp successfully.
INFO:analyzer:[4/4] clangsa analyzed lib.cpp successfully.
```
In this commit the logger will be initialized and a new --verbose option
will be introduced for the logger to debugging.