[analyze] Analyze headers with the analyze skip file
Multiple source files may depend on a header file. If the skip file
given to the analyze command contains a header file with this patch
CodeChecker will try to find all the dependencies of these headers
and extend the skip file with these source files by analyzing these
source files too.

It can be used to analyze only the changed files properly. The only
limitation is that the full compilation database is needed to collect
these information.