[CLI] Extending analyzer commands with checker config options
Checker config options from --checker-config flag are added to the
respective analysis commands (clangsa and clang-tidy).

Bugfix:
- Earlier the content of tidy config given by --tidy-config has been
added to the -config= flag of clang-tidy binary between extra quotation
marks. -config flag requires a yaml/json object, however with the extra
quotation mark it was a string object.
- clang-tidy checker config file has a yaml format that contains --- at
the beginning and ... at the end. These are page delimiters in the yaml
file. For some reason ... has been omitted before forwarding it to
-config= and this resulted a malformed yaml.