[analyzer] refactor how the analyzer configs are passed for each process
No sync manager is required anymore. The configs were already
shared in a read only way. A configuration change in one process
should not influence the config in another.
Every process will get its own local configuration instance in
the new setup and can modify it if needed.

Every configuration change which could influence the analysis
should be done before the worker pool starts the analysis jobs.

The refactoring was done to fix a hardly reproducible
problem with the sared configuration values between processes.

For some reason the socket belonging to the SyncManager was closed
by the SyncManager even if data was tried to be sent to be shared
between the processes.

  File "<string>", line 2, in get
  File ".../python3.8.2/multiprocessing/managers.py", line 844, in _callmethod
    conn.send((self._id, methodname, args, kwds))
  File ".../python3.8.2/multiprocessing/connection.py", line 215, in send
    self._send_bytes(_ForkingPickler.dumps(obj))
  File ".../python3.8.2/multiprocessing/connection.py", line 434, in _send_bytes
    self._send(header + buf)
  File ".../python3.8.2/multiprocessing/connection.py", line 380, in _send
    n = write(self._handle, buf)
TypeError: an integer is required (got type NoneType)

One other thing which could cause this problem could have been the
shares SyncManager between two multiprocessing pools (pre analysis,
and the normal analysis), but that was not validated.