[server] Fix getting run lock
If there is already a lock in the database for the given run name which
is already expired and multiple processes are trying to get the lock
entry from the database for update we may get the following exception:
```
could not obtain lock on row in relation "run_locks"
```

Steps to reproduce:
- Put 5 seconds sleep before and after getting run lock entry from the
database in the `__store_run_lock` function.
- Make sure that there is already a run lock in the database for the
given run and it is expired.
- Run two storages which will update the same run.
- Exception will be thrown.