[analyzer] Always generate valid log file
It is possible that the log file will not be created or it will be empty
for example when the build command is an empty string or when there is no
compiler command to log. For this reason we will create this log file if
it does not exist at the beginning of the logging phase and we will insert
an empty array to it, to make sure that it will be a valid json file.