[analyzer] Exit with an error code if the logfile does not exist
- If the log file does not exist we will print an error message and
exit with an error code.
- If multiple log files were given to the CodeChecker analyze command,
we print an error message that it is not supported yet and we exit
with an error code. Instead of this we can just remove the multiple
log files support from the argument parser so the user will not be
able to give multiple log files for this command.