[web] Fix autologin when session expired
If auto login is configured by using username and password but the session
token is expired the client will not login the user automatically just print
the following error message:
```
Internal server error: Error code 401: Unauthorized!
```

This patch will solve this problem on the following way: if the client
call an API request with an expired session token, we will catch the exception,
try to login the user and call the API function again.

With this solution no unnecessary API request will be made to the server if
the session token is not expired.