[license] Change license header in source files
The source code right now is licensed under the lesser known
`University of Illinois/NCSA Open Source License (NCSA)`. In the last year
the LLVM project including Clang and Clang-tidy went through a re-licensing
process. The license from NCSA license was changed to
`Apache 2.0 with LLVM Exception`.

The CodeChecker project is closely related to these projects and after
some discussion in the team we decided to change the license of the
CodeChecker project to the better known "Apache 2.0 with LLVM Exception"
license.

This commit changes the license header in each source file.