[analyzer] Create reports on compiler failure only in case of Clang Tidy
In case of compiler errors Clang Tidy analyzer will create reports
(plist files) from the standard output. Previously if we enabled the
Clang SA analyzer too and we used one of the context free hashes we
tried to open the plist file and override the report hashes in it
but no plist file was created by this analyzer. So an exception was
thrown that the plist file does not exist. This commit will handle
this use case too.