[server] Improvement in a slow query preventing timeout
Earlier the "CodeChecker cmd diff..." command assembled an SQL query which
contained case insensitive pattern matching on the bug hashes with potential
joker characters. The number of these pattern matches is proportional to the
number of reports. This query could cause a timeout in the underlying SQL
server. The solution is checking equality instead of pattern matching in case
an exact report hash is used. If the hash contains a * joker character then we
fallback the original case insensitive pattern matching.