[analyzer] Fix analyzer --file option
When `--file` option was given to the CodeChecker analyze or check commands
at the backround a skip handler was created which included only the given
source file and skipped everything. The problem was when we found reports
in header files we skipped these reports too.
With this patch we will not skip results from the header files.

We will keep results in all the header files which have the following extenstions:
```
file.hh
file.H
file.hp
file.hxx
file.hpp
file.HPP
file.h++
file.tcc
```

For mor information see:
https://gcc.gnu.org/onlinedocs/gcc-10.1.0/gcc/Overall-Options.html#index-file-name-suffix-71