[cmd][web] improve storage performance
The speedup is done mainly on the client side after some refactoring:
 - parsing the plist files is done by a process pool in parallel,
   using all the cpu cores on the machine
 - the source files are checked for codechecker review comments
   with a process poll using all the available cores
 - a quick fast check (with a regex) was done to check if there
   are any review comments in a source file and do the more
   time consuming parsing in parallel if there is any
   (this change was done on the server side too)
 - every source file mentioned in the reports should be
   opened only once to check for review comments

The error handling was a bit changed to easier propagate back
the spelling errors.

After some measurements the speedup can be seen if there are
more than ~2000 reports, below that report count there is a
little slowdown because of the extra process poll creation
and management overhead.