[server] Filter out deduplicates from metadata.json
Mutliple reports directory can be given for the CodeChecker store command.
We will get metadata.json files from each report directory, we will merge
them together. The problem is that on the server we will put all the
version information to a list not a set so if all the report directory
were analyzed with the same CodeChecker version we will show the same version
multiple times.
This patch will use a set to filter out such duplicates in the version and
in the check command too.