[web] Draw bug path on viewport change
If the source file was too large (>1000) and the bug path was too long
(for example first bug step was at line 7 and the last bug step was at line 950)
some jsplumb lines which were not in the actual viewport were draw to a bad position.
This commit will not draw lines which are not in the viewport and will draw lines on
every viewport change.