[gui] Fix adding product with postgresql database
Database user name and password are not set properly when adding a new
product with PostgreSQL database. These values should be encoded to base64
on the client side before sending to the server. Also wrong field names were
used: username instead of username_b64 and password instead of password_b64.