[server][fix] Fix disconnection product after remove
When a product is removed but a server is still connects to this product
we will try to disconnecting it by removing from the local cache.
We will iterate over the `self.__products` dictionary keys and we will
remove this product from the dictionary. In Python 2 it worked but in
Python 3 the following exception was thrown:
```
RuntimeError: dictionary changed size during iteration
```
The problem is that in Python 2 the `keys` makes a copy of the keys
what we can iterate over while modifying the dictionary. In Python 3
`keys` returns an iterator instead of a list. To solve this problem
we will create a force copy from the keys.