move subcommands into bin/cmd in the package
With this change the subcommands are moved into a
sub directory under the bin directory so they are
not available in the PATH if the bin directory is
added there, so they can not be called by accident.
The top level scripts contain some additional configuration
and setup without that setup the subcommands can not be
executed in some environments.