[db] Add index for content hash column on files table
If we use PostgreSQL database and we set the query timeout to 2 minutes the
database cleanup will be timed out when the server tries to remove unused
files and the database is to big.
The reason is that the database sequentially scanned the database to look
for foreign key constraints. To solve this problem we add index to the
content hash column on files table.