[analyzer] collect implicit include paths in case of clang mismatch
If different clang binaries are used for compilation and analyisis
collect the implicit include paths from the clang doing the
compilation which can be used by the clang doing the analysis.

It is possible that the clang doing the compilation was compiled
differently (using std lib headers from a non standard place)
compared to the clang which is doing the analysis.
In this case the clang executing the analysis tries to use
the system headers from the host which might not be compatible
with the project which was built.