[gui] Show filters in expansion panels
We already have many report filters on the UI and it is hard to see
through where is the filter what the user actually want to set. For
this reason with this patch we will show each filter in an expansion
panel and every filter will be closed by default. We will also show
the selected filter items in the filter header bar which will be trimmed
when too many filter items are selected.