[analyzer] config file option for the parse subcommand
Add support to use a config file by the parse subcommand.
The subcommand arguments can be saved and used from the
CodeChecker config file.

Common config file checking and parsing parts were
moved to the codechecker_common module because
the analyzer and the server parts use the same code.

The config file 'analyze' command related config
key was 'analyzer' to keep it in sync with the subcommands
it is renamed to 'analyze' in the documentation and examples.
The code was made backward compatible to accept both cases
and give a warning if both sections are available
in the configuration file.

The subocommand configuration examples were updated
and extended.