[refactor] output formatters, type hints, report types
A new common output module was created for the output formats.
The output formatters should depend on the common Report type
and not use the thrift ReportData type so they can be shared
between multiple parts of the source code.

There is one case in the server command line client
where the json format uses the ReportData and that is
converted to json. For backward compatibility that is
kept for now.

The common module can be extended with additional output formats
later so they can be shared between the analyzer and the web parts.

The twodim format functions were moved that module too together
with the codeclimate and gerrit formatters.

Additionally the cmd_line_client file was refactored.
Multiple smaller functions were factored out not
to depend on function internal variables where
they were. It is easier to test them like that.

A new report type converter was added for the
type conversion between the Report and ReportData
types. Should be needed only at the web part.
The common and analyzer parts use only the Report type.

Type hints were added to multiple functions for
better understanding what type is expected where.

The plist parser module was changed because the
comments and the actual source code assumed different types.
Instead of the list of files a dict of files is used there.
The order is important! The key is the index of the source
file where it was in the list of files in the plist report.

The main section in the Report type does not reference the
source file as an index anymore to simplify the usage.
The source file path is stored there.
For the bug path sections this replacement was not done.