[gui] Remove regex icon from the report filters
Some of the report filters support custom search with * quantifiers.
If the user types something to the input box then a custom filter item
is appeared with a regex icon at the top of the filter items which allows the user
to filter by wildcard pattern. This is actually not a regex expression so
the regex icon is a little bit misleading. To solve this problem this patch
will remove this icon.