[gui] Keepalive component states
Use vue `keep-alive` built-in component to cache component states.

So for example when navigating from the Runs page to the Reports page
the state of the Runs page component is saved. So once we go back to the
Runs page component the user will see the previous state.

This will also help us to reduce API requests to the server.