Don't include gcc -o output in .plist hash
In case of "make 4.3" depending on compile-time options "make" tool
can be built so a subprocess named cc1build will be logged by
"CodeChecker log".
See posix_spawn() option:
https://lists.gnu.org/archive/html/info-gnu/2020-01/msg00004.html
In this case the -o output argument of cc1build command is a randomly
named temporary file. We can't afford dynamic parts in the original
build command, because its hash is used for identification in the
plist file name.