[fix] Wait while filters are initalized and cleared properly
- `initByUrl` is an async function which returns a Promise. Unfortunatelly
in the Run Filter it is called an async function (setSelectedItems) but we
forgot to wait for it. This way the filters are not initalized properly.
For example when we opened a product and clicked on a run name the reports
are not filtered by run name.
- `clear` functions can be async functions so we need to wait for them
before we update the report filters when clearing all reports filter.