[test] Fix detection date filter test
At the begining of the detection date filter tests we get one report from
the database based on the first two runs. The order of the result was
unspecified on the API call so the default order was by severity.
It was possible that on the first test run the report related to the
first run was queried from the database but on the second test run the
report related to the second run was queried. So this cause a
non-deteministic test run.