[gui] Do not keep url filters when navigating from run name/run history links
It is possible that the user compared two runs, went back to the run list page
and opened a simple run. In this case it is not soo good if we keep the previous
urls (compared to run names etc). So for this reason we keep url states only
if the user navigates between header menu items.