[analyzer] parse the compilation tag separately
The compilation target which could be different from the default
compiler target was not correctly parsed from the compilation command.
This could cause problems in case the project was cross compiled
to a different target.
The default compilation target is collected but that might be
different from the target used in the compile command.