[server] Set analyzer name for clang-diagnostic checkers
Checkers that start with `clang-diagnostic-` are compiler warnings and these
are not in the metadata.json file when storing results. We didn't handle this
use case so we were not able to set analyzer names for these checkers previously.
This commits solves this problem and set the analyzer name field for these
checkers too.