[server] ignore encoding errors for source file content
If for some reason a file content was stored with a content
which can not be decoded to utf-8 ignore those characters.

New tests were added where a source file was saved with latin1
encoding to see if the file content can be loaded back
without any execeptions.

The characters which can not be encoded are ignored.