[server][auth] user dn postfix preference
When searching for a user in the LDAP directory it is possible that
multiple DNs are found for a user. To select a prefered DN the
`user_dn_postfix_preference` value can be set at the server config and
it will return the first DN where the end of the DN matches
the user_dn_postfix_preference value.