[cmd] Catch and print every exceptions in the main script
Previously we handled the use case when no subcommand is given by catching
an Attribute exception. We cought this exception in our main script and
we print the help message of the CodeChecker command.
This is bad because the CodeChecker code can thrown this type of exception
a couple of ways. For example if we call a function on a NoneType variable.

For this reason we will print the help message of the given command only
if no subcommand is given.