[cmd] Base client helper
In the thrift_call.py file the `self._reset_token` function was undefined
in some cases for example when we tried to add a new product and the token
is expired. To solve this problem permanently in this patch we create a
base client helper class and use this when creating other client helper
classes. And this class will contain this reset_token method.