[e2e] Pass environment variables for every command
When we run the e2e tests and we initalize the database we run some
CodeChecker command (login, add product, store). Previously we set the
`CC_PASS_FILE` and `CC_SESSION_FILE` only for the login command so
the user was able to login automatically. In this commit we will set
this for every CodeChecker command call.