[gui] Improve component statistics page load
Use `getSeverityCounts` and `getReviewStatusCounts` API functions to get
severity and review status statistics for a component instead of calling
separate API functions for each column.

For example in case of 1 server setup it will reduce the time to fetch data
from ~24668 milliseconds to ~8311 milliseconds.

This way the Checker statistics and the Product overiew pages will load much
faster than before.