[gui] Set z-index of the last month input field to a lower value
Previously the z-index CSS option for the last month selector was set
to a really high number. This way if the user opened for example the
Number of failed files dialog, the last month input field was shown above
the dialog window.
This patch will set this CSS option to a lower value.