[gui] Fix links in statistics tables
If we are for example in the checker statistics page and we set some
filter (run, checker name) and we click on a link we need to keep
the previous run/checker name filters to filter the report.

So all filters set in the statistics view should follow to the Reports view.