[server] Fix exception when groupPattern is null
When groupPattern is null in server_config.json, the following
exception was being logged when a user tried to authenticate:

    File ".../codechecker_server/auth/cc_ldap.py", line 390, in get_groups
      group_pattern = group_pattern.replace('$USERDN$', user_dn)
  AttributeError: 'NoneType' object has no attribute 'replace'

groupPattern is also null in one of the examples in authentication.md.
In this case, the get() with default value that was used:

  ldap_config.get('groupPattern', '')

actually returned None, since the key actually existed in the dictionary
with value None.