[fix][server] review comment parsing with trim path
Fixes the source file paths extracted at the server and
used for review status parsing when the trim path
feature was used.
The server depends on the client to send all the source files
containing review status comments. After extraction the
received source files will be parsed for review comments
to update them in the database.
Because the wrong path was used the server did not found
the source file to parse.
The server did not throw any error if a source file was missing
because it is a valid usecase that the client does not send
all the source files because they are already available in the
database at the server side.

New test was added where the review comments were modified
between the two store commands (one with trim path) and then
the successful review comment updates are checked.