[server] Verify TLS certificate in LDAPS connections
If LDAP authentication is used in the server and the LDAP server
connection is using LDAPS, then reject the connection if the certificate
is not trusted by the system.

This can be overriden by setting the
tls_require_cert="never" in the configuration.

NOTE: The option used before to skip the SSL certificate verification
was OPT_X_TLS_ALLOW, which is documented in python-ldap as "used
internally by slapd server". Therefore it was replaced by
OPT_X_TLS_NEVER.