[analyzer] Fix skipping report from plist
Lets suppose that we are running the CodeChecker analyze command by using
multi cores and we are also using a skip file.

If the compilation database contains two translation units which
different only in the `-o` parameter, it will produce the same plist file.
It is possible that we try to modify this file from 2 threads
to skip reports from header files in a non-thread safe way. This way
the plist files may be corrupted.

In this patch we will use portalocker to lock the file before touching
it to make sure that only 1 process is writing to this file in the same time.