[analyzer] Fix plist file path attribute
When an older version of clang was used where the report hash is not in
the plist we tried to ovveride the plist file and insert the report hash
to the plist file.

When we parsed a plist file, in the main diagnostics section the
`location.file` attribute was set to a file path instead of an index which
caused some unexpected behaviour.