[server] Change review status date only if necessary
Review status is a shared table among runs. When multiple runs
are stored in parallel, there may be a race condition in updating
review status fields. The most common reason of deadlocks is
changing only the date to current date. This condition checks if
something else is also changed other than dates.