[github] Update codeowners
Unfortunately @gyorb left our company and he is not working on this
project anymore. So for this reason we should remove him from the
code owners list. This way he will not get an automatic review request
for every pull requests and it will not confuse the open source users
why he doesn't review the PR's.