[client] Diff "resolved" reports of remote to local
1. Analyze a project to "reports" dir.
2. Store it in run named "hello"
3. CodeChecker cmd diff -b reports -n hello --resolved

This scenario shows all reports which is a bad behavior.

When server side fetches a set of reports from database, it uses a filter
object. If "reportHash" filter attribute is empty list, it means that no
report hash was selected for diffing, i.e. it takes all reports into
account.